+++ CHANGELOG.md
@@ -1,3 +1,5 @@
+* Fixed syntax mistake in `trace.TablePooStateChangeInfo` to `trace.TablePoolStateChangeInfo` 
+
 ## v3.26.9
 * Fixed bug with convert ydb value to `time.Duration` in `result.Scan[WithDefaults,Named]()`
 * Fixed bug with make ydb value from `time.Duration` in `types.IntervalValueFromDuration(d)`